requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
psycopg2==2.8.3
